+++ src/main/java/com/supermartijn642/rechiseled/RechiseledClient.java
@@ -4,13 +4,10 @@ import com.supermartijn642.core.TextComponents;
 import com.supermartijn642.core.gui.WidgetContainerScreen;
 import com.supermartijn642.core.registry.ClientRegistrationHandler;
 import com.supermartijn642.rechiseled.screen.BaseChiselingContainerScreen;
-import net.minecraftforge.api.distmarker.Dist;
-import net.minecraftforge.fml.common.Mod;
 
 /**
  * Created 21/12/2021 by SuperMartijn642
  */
-@Mod.EventBusSubscriber(value = Dist.CLIENT, bus = Mod.EventBusSubscriber.Bus.MOD)
 public class RechiseledClient {
 
     public static void register(){