+++ src/main/java/com/supermartijn642/rechiseled/screen/ScreenItemRender.java
@@ -1,8 +1,12 @@
 package com.supermartijn642.rechiseled.screen;
 
 import com.mojang.blaze3d.vertex.PoseStack;
-import com.supermartijn642.core.gui.ScreenUtils;
+import com.supermartijn642.core.ClientUtils;
+import com.supermartijn642.core.render.RenderUtils;
+import net.minecraft.client.renderer.LightTexture;
+import net.minecraft.client.renderer.texture.OverlayTexture;
 import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemDisplayContext;
 
 /**
  * Created 22/01/2022 by SuperMartijn642
@@ -14,13 +18,21 @@ public class ScreenItemRender {
      */
     public static void drawItem(PoseStack poseStack, Item item, double x, double y, double scale, float yaw, float pitch, boolean doShading){
         scale /= Math.sqrt(2 + 1d / (16 * 16));
-        scale /= 10;
+        scale *= 1.6;
 
         poseStack.pushPose();
         poseStack.translate(x, y, 0);
         poseStack.scale((float)scale, (float)scale, (float)scale);
-        poseStack.translate(-8, -8, 0);
-        ScreenUtils.drawItem(poseStack, item.getDefaultInstance(), null, 0, 0);
+        ClientUtils.getItemRenderer().renderStatic(
+            item.getDefaultInstance(),
+            ItemDisplayContext.GUI,
+            LightTexture.FULL_BRIGHT,
+            OverlayTexture.NO_OVERLAY,
+            poseStack,
+            RenderUtils.getMainBufferSource(),
+            null,
+            0
+        );
         poseStack.popPose();
     }
 }